fixes for MOB-21261, MOB-21262, and MOB-21447
MOB-21447: replaces the heads up notification (HUD) on API 21 with the expanded push template as the heads up notification is not expandable. additionally, the heads up notification is replaced by the collapsed layout on API22 and 23 as the HUD notification is expandable but does not show any title or body text.
MOB-21261: use the `setOnlyAlertOnce` on the NotificationCompat.Builder to prevent an interaction intent from creating a new notification each time a push template interaction occurs.
MOB-21262: use a expanded layout with a smaller image area to allow the notification action buttons to be displayed on the bottom of each notification.